created WorldManager, ItemSpawner (not fully functioning), item class script, 1 temp prefab
item class script now exists, a rock prefab has been made to test with (will be replaced with good prefabs once models are made). Added a WorldManager script to be used in the furutre to (re)set the environment. Created SpawnItems script for the spawning items code when it gets made.