World building (#22)
* Mounds and Improved Digging

Modified the TerrainEditor script to make holes grow in diameter the deeper they are dug and added a function to make a mound at a position.

* Added Environment Model

Imported the environment model and placed it relative to the unity terrain.

* Fixed Merge Issues Hopefully

Brought in changes and re-introduced the environment model with textures.

* Adjusted Item Integration

Added tags and assigned item prefabs to let them spawn in the world.

* Adjusted item spawning zone

Adjusted item spawning zone and made the digging action create a slightly bigger hole.

* Perlin Noise and Particle Changes

The sand is now textured by Perlin Noise and the particles have a texture. The texture is part of a collection from Kenney.nl.

* Adjustments to most current version

Adapted WorldBuilding progress to work with new item spawning. Additionally, adjusted particles and added alpha rocks to block the tunnel opening.