World building (#35)
* Mounds and Improved Digging

Modified the TerrainEditor script to make holes grow in diameter the deeper they are dug and added a function to make a mound at a position.

* Added Environment Model

Imported the environment model and placed it relative to the unity terrain.

* Fixed Merge Issues Hopefully

Brought in changes and re-introduced the environment model with textures.

* Adjusted Item Integration

Added tags and assigned item prefabs to let them spawn in the world.

* Adjusted item spawning zone

Adjusted item spawning zone and made the digging action create a slightly bigger hole.

* Perlin Noise and Particle Changes

The sand is now textured by Perlin Noise and the particles have a texture. The texture is part of a collection from Kenney.nl.

* Adjustments to most current version

Adapted WorldBuilding progress to work with new item spawning. Additionally, adjusted particles and added alpha rocks to block the tunnel opening.

* Added Colliders and Decoration

Added colliders to objects and defined the home area as a no-digging zone.

* Minor adjustment to castle collider

Fixed size of sand castle collider.

* Decorative assets and materials

Added the palm tree model and assigned textures to models for alpha.

* Claws have become independent

Claws are no longer children of the crab to enable rigid bodies on both.

* Commented out Debugs

Tested that everything is working and commented out debug.log statements.

* Added Active Claw Indication

Crab claws now raise and drop based on which is the active claw.