Modified textures and other small changes (#58)
* Mounds and Improved Digging

Modified the TerrainEditor script to make holes grow in diameter the deeper they are dug and added a function to make a mound at a position.

* Added Environment Model

Imported the environment model and placed it relative to the unity terrain.

* Fixed Merge Issues Hopefully

Brought in changes and re-introduced the environment model with textures.

* Adjusted Item Integration

Added tags and assigned item prefabs to let them spawn in the world.

* Adjusted item spawning zone

Adjusted item spawning zone and made the digging action create a slightly bigger hole.

* Perlin Noise and Particle Changes

The sand is now textured by Perlin Noise and the particles have a texture. The texture is part of a collection from Kenney.nl.

* Adjustments to most current version

Adapted WorldBuilding progress to work with new item spawning. Additionally, adjusted particles and added alpha rocks to block the tunnel opening.

* Added Colliders and Decoration

Added colliders to objects and defined the home area as a no-digging zone.

* Minor adjustment to castle collider

Fixed size of sand castle collider.

* Decorative assets and materials

Added the palm tree model and assigned textures to models for alpha.

* Claws have become independent

Claws are no longer children of the crab to enable rigid bodies on both.

* Commented out Debugs

Tested that everything is working and commented out debug.log statements.

* Added Active Claw Indication

Crab claws now raise and drop based on which is the active claw.

* Adjusted Dig

Lowered values to make digging slower

* Added Crab Rock

Added and positioned the crab rock model and created material to apply to it.

* Updated Textures

Updated palm tree and crab rock textures

* New spawning area and footsteps

Added the new spawning area on the other side of the map and added a preliminary footsteps particle effect.

* Adjustments for the new crab

Changed crab claw movement, changed item holding, made digging a little faster, adjusted particles for new crab, particles are now dynamic to react to analogue input, and maybe other things.

* Added models and textures for new castle, flag, and sign

New castle, flag, and sign

* Placed models and small adjustments

Placed models for the new sand castle, flag, and sign. Adjusted digging. Turned of visibility of TerrainManager.

* Adjusted textures

Modified textures for the environment, sand castle, and palm tree

* Turned off outline code for left claw

Left claw still had code that was breaking things.